improve nginx entrypoint

+++ rootfs/bin/nginx.sh
@@ -15,6 +15,7 @@ export SSL_CERTS_PATH=${SSL_CERTS_PATH:-"/certs"}
 export SSL_CERT_BASENAME="${SSL_CERT_BASENAME:-"cert"}"
 export SSL_CERT_PATH=${SSL_CERT_PATH:-"${SSL_CERTS_PATH}/$SSL_CERT_BASENAME.crt"}
 export SSL_KEY_PATH=${SSL_KEY_PATH:-"${SSL_CERTS_PATH}/$SSL_CERT_BASENAME.key"}
+export NGINX_INJECT_DIR="${NGINX_INJECT_DIR:/nginx.d}"
 export NGINX_HTTP_PROTECT_USER=${NGINX_HTTP_PROTECT_USER:-root}
 export NGINX_HTTP_PROTECT_PASSWORD=${NGINX_HTTP_PROTECT_PASSWORD-}
 export NGINX_SKIP_CHECK="${NGINX_SKIP_CHECK-}"
@@ -34,15 +35,29 @@ export NGINX_FREP_SKIP=${NGINX_FREP_SKIP:-"(\.skip|\.skipped)$"}
 export NGINX_SKIP_EXPOSE_HOST="${NGINX_SKIP_EXPOSE_HOST-}"
 export NGINX_DH_FILE="${NGINX_DH_FILE-/certs/dhparams.pem}"
 export NGINX_DH_FILES="$NGINX_DH_FILE"
+export NGINX_PASSWORDS_DIR="${NGINX_PASSWORDS_DIR:-/etc/htpasswd}"
+export NGINX_PASSWORD_FILE="${NGINX_PASSWORD_FILE:-${NGINX_PASSWORDS_DIR}/protect}"
 export NGINX_CONF_RENDER_DIR="${NGINX_CONF_RENDER_DIR:-"/tmp/nginxconf"}"
 # let cron wrapper script use the custom nginx conf
 export NO_NGINX_LOGROTATE=${NO_NGINX_LOGROTATE-}
-export NGINX_CONFIGS="${NGINX_CONFIGS-"$( \
-    find "$NGINX_CONF_DIR" -type f \
-    |grep -E -v "$NGINX_FREP_SKIP|\.template$")
+export VHOST_TEMPLATES="${VHOST_TEMPLATES-"/etc/nginx/conf.d/default.conf"}"
+export VHOST_TEMPLATE_EXTS="${VHOST_TEMPLATE_EXTS:-frep template}"
+export NGINX_CONFIGS="${NGINX_CONFIGS-"
+$(find "$NGINX_CONF_DIR" -type f |grep -E -v "$NGINX_FREP_SKIP|\.(${VHOST_TEMPLATE_EXTS// /|})$")
 /etc/logrotate.d/nginx"}"
+create_file() {
+    for i in $@;do
+        if [ ! -e "$(dirname $i)" ];then mkdir -p "$i";fi
+        touch "$i" && chown $NGINX_USER:$NGINX_USER "$i" && chmod 640 "$i"
+    done
+}
 log() { echo "$@" >&2; }
 vv() { log "$@";"$@"; }
+if [ "x${NGINX_USER}" = "x" ];then
+    for i in www-data nginx www root;do
+        if (getent passwd $i >/dev/null 2>&1);then NGINX_USER=$i;fi
+    done
+fi
 # patch rsyslog default conf not to interfer with nginx self contained
 for i in /etc/logrotate.d/rsyslog;do
     if [ -e $i ];then
@@ -50,16 +65,36 @@ for i in /etc/logrotate.d/rsyslog;do
         ( sed -i -re "/\/nginx\/|nginx.log|\/\*-(error|access)/ d" $i || true)
     fi
 done
-touch /etc/htpasswd-protect
-chmod 644 /etc/htpasswd-protect
-if [ "x$NGINX_HTTP_PROTECT_PASSWORD" != "x" ];then
-  echo "Activating htpasswd for $NGINX_HTTP_PROTECT_USER">&2
-  echo "$NGINX_HTTP_PROTECT_PASSWORD" \
-	  | htpasswd -bim /etc/htpasswd-protect "$NGINX_HTTP_PROTECT_USER"
-fi
+# We search for XXX_HTTP_PROTECT_PASSWORD/XXX_HTTP_PROTECT_USER/XXX_HTTP_PROTECT_FILE envvars
+# to generate relative htpasswd files
+# Note that $NGINX_PASSWORD_FILE will receive all defined password pairs
+OIFS=${IFS-}
+IFS=$'\n'
+for envline in $(env|grep -E "^([^\s ]+_)?HTTP_PROTECT_PASSWORD=");do
+    value="$(echo $envline|sed -re "s/^([^=]+)=(.*)/\2/g")"
+    variable="$(echo $envline|sed -re "s/^([^=]+)=(.*)/\1/g")"
+    if [ "x${value}" != "x" ];then
+        uservariable="$(echo $variable|sed -re "s/PASSWORD$/USER/g")"
+        filevariable="$(echo $variable|sed -re "s/PASSWORD$/FILE/g")"
+        prefix="$(echo "$variable"|tr '[:upper:]' '[:lower:]'|sed -re "s/_http_protect_password//gi")"
+        default_passwdfile="$NGINX_PASSWORDS_DIR/${prefix}protect"
+        echo $filevariable
+        passwd_file="$(eval "echo "\${${filevariable}:-\$default_passwdfile}"")"
+        passwd_user="$(eval "echo "\${${uservariable}:-root}"")"
+        log "Generating HTPASSWD for $uservariable ($passwd_file & $NGINX_PASSWORD_FILE)"
+        for i in "$passwd_file" "$NGINX_PASSWORD_FILE";do
+            create_file "$i" && echo "$password"| htpasswd -bim "$i" "$passwd_user"
+        done
+    fi
+done
+IFS=${OIFS}
+# retrocompat: link all subpassword files in /etc/htpasswd/* to /etc/htpasswd-$i counterparts
+for i in $(find "$NGINX_PASSWORDS_DIR" -type f -maxdepth 1);do ln -sfv "$i" "/etc/htpasswd-$(basename $i)";done
+###
 if [[ -z ${NGINX_SKIP_EXPOSE_HOST} ]];then
-    ip -4 route list match 0/0 \
-        | awk '{print $3" host.docker.internal"}' >> /etc/hosts
+    if ( ip -4 route list match 0/0 >/dev/null 2>&1 );then
+        ip -4 route list match 0/0 | awk '{print $3" host.docker.internal"}' >> /etc/hosts
+    fi
 fi
 if [ "x$NGINX_STD_OUTPUT" = "x" ] && [ "x$NGINX_AUTOCLEANUP_LOGS" != "x" ];then
     for i in $NGINX_LOGS_DIR;do rm -fv $NGINX_LOGS_DIR/*;done
@@ -69,16 +104,19 @@ for e in $NGINX_LOGS_DIRS $NGINX_CONF_DIR;do
     if [ "x$NO_CHOWN" != "x" ] && [ -e "$e" ];then chown "$NGINX_USER" "$e";fi
 done
 if [ "x$SKIP_CONF_RENDER" = "x" ];then
-    for i in $NGINX_CONFIGS;do frep $i:$i --overwrite;done
+    for i in $NGINX_CONFIGS;do frep "$i:$i" --overwrite;done
+    for v in ${VHOST_TEMPLATES};do
+        for e in ${VHOST_TEMPLATE_EXTS};do
+            if [ -e $v.$e ];then frep "${v}.$e:$v" --overwrite;fi
+        done
+    done
 fi
 # also render an eventual /nginx.d folder
-if [ "x$SKIP_EXTRA_CONF" = "x" ] && [ -e /nginx.d ];then
+if [ "x$SKIP_EXTRA_CONF" = "x" ] && [ -e "$NGINX_INJECT_DIR" ];then
     log "Nginx conf injection directory present, processing"
-    if [ -e "$NGINX_CONF_RENDER_DIR" ];then
-        rm -rf "$NGINX_CONF_RENDER_DIR"
-    fi
+    if [ -e "$NGINX_CONF_RENDER_DIR" ];then rm -rf "$NGINX_CONF_RENDER_DIR";fi
     mkdir -p "$NGINX_CONF_RENDER_DIR"
-    cp -rf /nginx.d/* "$NGINX_CONF_RENDER_DIR"
+    cp -rf "$NGINX_INJECT_DIR"/* "$NGINX_CONF_RENDER_DIR"
     if [ "x$SKIP_CONF_RENDER" = "x" ];then
         for v in $(cd "$NGINX_CONF_RENDER_DIR" && find . -type f);do
             if (echo "$f" |grep -E -v "$NGINX_FREP_SKIP");then
@@ -110,7 +148,7 @@ else
             done
         else
             nginxconfs="$(find /etc/nginx/ -type f|xargs cat)"
-            if [ "x$nginxconfs" != "x0" ];then
+            if [ "x$(echo "$nginxconfs"|wc -l)" != "x0" ];then
                 for i in $( echo "$nginxconfs"\
                     | grep -E "\s*ssl_dhparam" | grep -v "{{" \
                     | awk '{print $2}'|sed -re "s/;//g"|awk '!seen[$0]++' );do
@@ -166,4 +204,4 @@ if [ "x$NGINX_SKIP_CHECK" = "x" ];then
 fi
 set -x
 exec $NGINX_BIN "$@"
-# vim:set et sts=4 ts=4 tw=80:
+# vim:set et sts=4 ts=4 tw=0:
